use a die directly in the @INC hook
Using a die directly in the @INC hook matches perl's behavior for a
missing module closer than using a fake module that dies.  It is also
simpler and more compatible.

On perl 5.8, with a die inside a module, the %INC entry will be left
behind.  Later attempts to load the module will then appear to succeed.
Putting the die directly will leave the %INC entry non-existent, just
like if the file does not exist.

This also avoids PerlIO, so it works on perl 5.6 without any extra work.

The error message was also updated to include " in @INC", so it is
closer to the error provided by perl.  This is expected by some modules
handling optional loads, such as Class::Load.